refactor(git-auth, github-api): consolidate token resolution and merge GH clients
Three near-identical implementations of the per-user-secret-then-env
token resolution existed across github-pr.ts, github-actions.ts, and
git.ts. github-actions.ts's copy carried the comment 'Mirrors the
resolution in github-pr.ts' — exactly the drift risk the architecture
audit flagged.

Two changes in one PR:

1. Extract resolveProviderToken() into src/git-auth.ts. One implementation,
   one set of test cases. Callers across github-pr/actions/git.ts now use it.

2. Merge github-pr.ts + github-actions.ts into github-api.ts. Both were
   GitHub REST clients with duplicated header builders, auth, and parseGithubRepo.
   github-actions.ts already imported parseGithubRepo from github-pr.ts —
   the split was artificial.

Security-adjacent: the look-alike-host rejection (hosts.ts) still applies
through resolveProviderToken's host validation. test/dodo.test.ts:987
('rejects look-alike github.com hosts') passes unchanged.
