+++ src/mcp.ts
@@ -4,8 +4,7 @@ import { z } from "zod";
 import { getSharedIndexStub, getUserControlStub, isAdmin, resolveAdminEmail } from "./auth";
 import { log } from "./logger";
 import { messageLimiter, promptLimiter } from "./rate-limit";
-import { createDraftPrForRun, createGithubRepo } from "./github-pr";
-import { pollVerifyWorkflow, triggerVerifyWorkflow } from "./github-actions";
+import { createDraftPrForRun, createGithubRepo, pollVerifyWorkflow, triggerVerifyWorkflow } from "./github-api";
 import { getKnownRepo, listKnownRepos } from "./repos";
 import { forkSessionInternal, SourceSessionMissingError } from "./sessions";
 import { errorResult, mcpUserEmail, propagateMcpDepth } from "./mcp-shared";
